+++ scm/main/src/core/mod.rs
@@ -1,5 +1,2 @@
 pub(crate) mod noop;
 pub(crate) mod processor;
-
-#[cfg(any(feature = "postgres", feature = "sqlite"))]
-pub(crate) mod refinery;